refactor: extract progressiveRecompressor to ingest_recompressor.go
Move progressiveRecompressor struct and all its methods (newProgressiveRecompressor,
MarkDone, Wait, runCompression, compressTableChunks) to a dedicated file.
Rename ingest_backfill_test.go → ingest_recompressor_test.go since it only
tests recompressor watermark logic.

Zero behavior change — pure file reorganization.
